Mod tests for JDK 1.1, arbibrary log4j, second try
git-svn-id: https://svn.eu.apache.org/repos/asf/logging/log4j/branches/v1_2-branch@311502 13f79535-47bb-0310-9956-ffa450edef68

+++ tests/src/java/org/apache/log4j/FileAppenderTest.java
@@ -32,16 +32,20 @@ public class FileAppenderTest extends TestCase {
    *
    */
   public void testDirectoryCreation() {
-      File newFile = new File("output/newdir/temp.log");
-      newFile.delete();
-      File newDir = new File("output/newdir");
-      newDir.delete();
+      //
+      //   known to fail on JDK 1.1
+      if (!System.getProperty("java.version").startsWith("1.1.")) {
+          File newFile = new File("output/newdir/temp.log");
+          newFile.delete();
+          File newDir = new File("output/newdir");
+          newDir.delete();
 
-      org.apache.log4j.FileAppender wa = new org.apache.log4j.FileAppender();
-      wa.setFile("output/newdir/temp.log");
-      wa.setLayout(new PatternLayout("%m%n"));
-      wa.activateOptions();
+          org.apache.log4j.FileAppender wa = new org.apache.log4j.FileAppender();
+          wa.setFile("output/newdir/temp.log");
+          wa.setLayout(new PatternLayout("%m%n"));
+          wa.activateOptions();
 
-      assertTrue(new File("output/newdir/temp.log").exists());
+          assertTrue(new File("output/newdir/temp.log").exists());
+      }
   }
 }
